Patch "contentItems" to no longer return trashed content.
The ``contentItems`` is used for other methods such as
``listFolderContents`` and ``getFolderContents``.
We do not want those methods to include trashed content as they are
used within Plone and within our addons.
We are not patching the underlying ``objectItems`` though as it
probably would make the trashed content disappear from ZMI and harder
to restore etc.