Fix user action name (from "user_management" to "trash").
The name "user_management" was a copy / pasted from
ftw.usermanagement. That means when both were installed, only the
action of the one installed last was displayed.

This only fixes the action on the ftw.trash side. If
ftw.usermanagement was also installed, the user-management-action may
be missing. Consider reinstalling ftw.usermanagement in this case.